`data.utils` -> `data.splitting`
