+++ chemprop/v2/cli/predict.py
@@ -11,7 +11,7 @@ from lightning.pytorch.callbacks import ModelCheckpoint, EarlyStopping
 import torch
 
 from chemprop.v2 import data
-from chemprop.v2.data.utils import split_data
+from chemprop.v2.data.splitting import split_data
 from chemprop.v2.models import MetricRegistry, modules
 from chemprop.v2.featurizers.reaction import RxnMode
 from chemprop.v2.models.loss import LossFunction, LossFunctionRegistry
